Add BasicApiHandler, a default crud/watcher resources implementation
Refactor the on-demand/capture in order to leverage the BasicApiHanlder
Split out alert from the graph. The alert manager and the websocket part
of alert are now separated. The new alert manager now relies on the
BasicApiHandler and leverage the watcher mechanism.
Add a mechanism to stop properly the api handlers.

Fix some races when stopping graph server and the sflow agent.

Change-Id: I8d558f5e042ad56b3293b26878d41640d45e6bd8