Split out the rendering function of flow table
This patch splits out the rendering functions of flow
tables and move them close to the API as they are related
to the way of providing informations about the flow tables
and not the way of the flow tables are working. Moreover this
patch refactors the UTs introducing testhelpers for flow tables
that will be used to add more UTs on the API side.

This patch also removed some test around collision as there was no
added values on these tests which were basically testing the go map.

Change-Id: I80a2896573cc976183f68d802edad65e59ca3b37