ovsdb: Add external_ids to metadata.
There are useful bits of information stored in the external_ids column
for an OVS interface.  For example, the "attached-mac" is the MAC
address of the device attached to this interface.  The "attached-mac"
is the MAC address that is seen in the VM and is what must be used to
correlate with a Neutron port.  The "iface-id" is also useful, as it
provides the Neutron port UUID in the Neutron case.

Details of these well-known external-ids can be found in:
https://github.com/openvswitch/ovs/blob/master/IntegrationGuide.md

Change-Id: I1653709965cc4f95666a6ac948e143ad6b76c68d
Signed-off-by: Russell Bryant <rbryant@redhat.com>