flow: add FlowUUID to flow and sflow agent allocator
This patch introduces a FlowUUID which is the real flow uuid
and add probepath to compute the UUID of a flow. This UUID
is used in flow table as unique ID. Having a unique ID per
probepath helps to follow packet going throught probes.
Due to the fact that the probepath is used to compute the
UUID it needs to be provided in a sync way instead of the
previous way which was async that why this patch change the
behavior of the ovssflow probe by introducing a sflow agent
allocator which helps to get the probepath without using
any graph lookup to find it.

Change-Id: I26db5aaa660e307a0d0ba285b0a37453f013c0f1