graph: fix non ovs internal added first by ovs
A non internal ovs interface can be fully qualified
and added by ovsdb before netlink in that case
two interfaces were created in the graph. This
patch checks in netlink if there if an interface
with the same ifindex and name has a UUID which
means it's the same interface.

Change-Id: Icf211b84c2e7d1fc41729acd66c042eff36874d3